Consolidate conversion helpers into OTPToken.swift
